снова вернул sizet для строк и колон. Нужно сделать считывание клеток из строки

+++ lab3/src/EnterParametersFromConsole.cpp
@@ -12,7 +12,7 @@ void EnterParametersFromConsole::initialFieldParameters(Field &field) {
 
 void EnterParametersFromConsole::initialFieldSize(Field &field) {
     ExceptionHandling exceptionHandling;
-    long long rows, columns;
+    size_t rows, columns;
     string input;
 
     cout << "Enter quantity rows:";
